feat: add release messaging
