+++ README.md
@@ -31,6 +31,7 @@ YOUR_SLACK_WEBHOOK: Webhook URL from Slack Incoming Webhook application
 
 <!-- start usage -->
 <!-- Warning: Content between these comments is auto-generated. Do NOT manually edit. -->
+
 ```yaml
 - uses: reside-eng/workflow-status-notification-action@v1
   with:
@@ -45,11 +46,18 @@ YOUR_SLACK_WEBHOOK: Webhook URL from Slack Incoming Webhook application
     # Required: true
     slack-webhook: ''
 
+    # Whether build is a release build (merge-main or select-for-release workflows)
+    #
+    # Default: false
+    # Required: true
+    is-release: ''
+
     # Status of the current run
     #
     # Required: true
     github-token: ''
 ```
+
 <!-- end usage -->
 
 ## Examples
