+++ src/inputs.ts
@@ -1,0 +1,6 @@
+export enum Inputs {
+  CurrentStatus = 'current-status',
+  SlackWebhook = 'slack-webhook',
+  GithubToken = 'github-token',
+  IsRelease = 'is-release',
+}
